(PUP-7946) Fix test requiring Ruby 2, do not load when 1.9.3
